much nicer implementation!
